components/auth_service_app/pyproject.toml
[tool.poetry]

name = "Auth Service API"
version = "0.1.0"
description = "Auth Service API"
authors = ["Dilip Sharma"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.14"  # Python runtime version
pydantic-settings = "^2.7.1"  # Settings management using Pydantic
# TEMP: Use development branch for instrumentation-hub-fastapi
instrumentation-hub-fastapi = { git = "https://github.com/vyavasthita/instrumentation-hub.git", rev = "master", subdirectory = "packages/python/fastapi" }
sqlalchemy = "^2.0.45"
aiomysql = "^0.3.2"
greenlet = "^3.0"
email-validator = "^2.2.0"
bcrypt = "^5.0.0"
pyjwt = "^2.10.1"

[tool.poetry.group.core.dependencies]
fastapi = "^0.115.8"  # FastAPI web framework
uvicorn = "^0.34.0"  # ASGI server for FastAPI


[tool.poetry.group.test.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
coverage = "^7.13.1"
anyio = "^4.12.1"
httpx = "^0.28.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"